requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
